refactor: update import paths to use data types from the data folder
